I forgot to add comments to the first comment. The first commit I added the basic login page code for Vue <script> and <template>.  No styling.  I added routes to the index.js and modified the APP.vue file to add the Login to the list selection.
This second commit was purely working with TailwindCSS to get the styling in line with the other pages.  Along with getting the input fields aligned veritically with good spacing.  I have a placeholder for the API call to get the userid and password from the DB.

No encryption is set up yet.  Verifying with the professor if this is needed for Sprint 2.  For now I am going to move onto the next requirement of Sprint 2.